+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 1.x branch
 ## 1.1 branch
+### 1.2.1
+* updated for `crabstudio/recaptcha` 2.1.1.
+
+### 1.2.0
+* migrated to CakePHP 3.6.
+
 ### 1.1.1
 * updated for CakePHP 3.6;
 * now it uses the `mirko-pagliai/php-tools` package. This also replaces